reorganized files to separate dimensions
