+++ converter/lookml/__init__.py
@@ -1,0 +1,1 @@
+from lookml import lookml_to_semantic_manifest
